Reorganize pages and give each service a page
This commit makes two major changes. The first is to make use of
some recent Jekyll features to better organize our metadata and
make it easier to modify the site in the future.

The main result of this is that we can get rid of the `_data`
folder and instead put all of the metadata in individual MD pages.
This allow many simplifications, including removing some keys that
were only there to ensure that the data YML files and MD pages
were properly linked.

Unfortunately, once I did all of this simplification, I realized
that one of the features that I was relying on to simplify things
is not actually available to us yet. Specifically, Jekyll v4.0
understands an `order` key in a collection's metadata, which will
make the order of that collection fixed when you do something like

    {% for product in site.products %}
    ...
    {% endfor %}

Since Github Pages will eventually update to Jekyll v4.0, I wanted
to keep the `order` key in `_config.yml`. Fortunately it is possible
to read the value of that despite Jekyll v3.8.5 not doing anything
for it natively. So, all of the for loops above have been augmented
to something like

    {% assign products = site.collections
      | where: 'label', 'products'
      | first %}
    {% for ordered in products.order
    {% for product in site.products %}
    {% if product.path contains ordered %}
    ...
    {% endif %}
    {% endfor %}
    {% endfor %}

While this is a pretty verbose and definitely inefficient replacement,
it still seems like the best solution because it's temporary.
When Github Pages updates to support Jekyll v4.0 we can simply remove
all the lines above to match the first snippet above.

The second major change is to give each service a separate page.
This page is pretty simple, only contains a two-column layout
where one column is the image and one is the text, and then the
contact form with the subject pre-filled in. The listing page
had to be changed to be more similar to the All Products page;
since they were then basically copies of one another, I generalized
them to a single layout that they both use.